Move DiskInterfaceTest into its own source file.
TEST=ninja_test --gtest_filter=DiskInterface*
